refactor the EditDocumentModal component
